ServerCache-18: Moved 'BinDiffClient' to separate class so that it can be later used inside tests
